move copy link confirmation view to its own file
